Remove blank lines from html
I found out Liquid adds a bunch of whitespace on every template {% %}
There is a built in way to remove it {%- -%} but this requires editing
every usage. This layout does the same without a huge hassle.
It shouldn't affect the final page but it makes life slightly nicer
for anyone who happens to view source. Or make edits and then diff
the old vs new html.